suppression/update images ok

+++ controllers/sauce.js
@@ -1,4 +1,5 @@
 const Sauce = require('../models/sauce');
+const fs = require('fs');
 
 //Créer une sauce 
 exports.createSauce = (req, res, next) => {
@@ -23,16 +24,47 @@ exports.getOneSauce = (req, res, next) => {
 
 //Modifier une sauce
 exports.updateSauce = (req, res, next) => {
-    Sauce.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet modifié !'}))
+    // on vérifie s'il y a un champ file dans la requete -> Si ? sinon :
+    const objetSauce = req.file ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : {...req.body };
+    //on supprimer le userId de la requete par mesure de sécurité
+    delete objetSauce.userId;
+    Sauce.findOne({ _id: req.params.id })
+        .then((sauce) => {
+            if (sauce.userId != req.auth.userId){
+                res.status(401).json({ message: 'Non autorisé !'})
+            }else{
+                //Suppression de l'ancienne image avant de MaJ la sauce 
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {});
+                //
+                Sauce.updateOne({ _id: req.params.id }, { ...objetSauce, _id: req.params.id })
+                    .then(() => {res.status(200).json({message: 'objet modifié!'})})
+                    .catch(error => res.status(400).json({ error }));
+            }
+        })
         .catch(error => res.status(400).json({ error }));
 };
 
 //Supprimer une sauce
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({_id: req.params.id})
-        .then(() => {res.status(200).json({message: 'Deleted!'})})
-        .catch((error) => {res.status(400).json({error})});
+    Sauce.findOne({ _id: req.params.id})
+       .then(sauce => {
+           if (sauce.userId != req.auth.userId) {
+               res.status(401).json({message: 'Non autorisé !'});
+           } else {
+               const filename = sauce.imageUrl.split('/images/')[1];
+               fs.unlink(`images/${filename}`, () => {
+                   Sauce.deleteOne({_id: req.params.id})
+                       .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                       .catch(error => res.status(401).json({ error }));
+               });
+           }
+       })
+       .catch( error => {res.status(500).json({ error });
+       });
 };
 
 //Afficher toutes les sauces
